requirements.txt
aiogram==3.4.1
python-dotenv
psycopg2-binary
flask==3.1.1
